Auth

+++ backend/src/main/java/com/swekit/backend/Service/SmsService.java
@@ -2,6 +2,7 @@ package com.swekit.backend.Service;
 
 import com.swekit.backend.Model.Sms;
 import com.swekit.backend.Model.SmsRequest;
+import com.swekit.backend.Model.User;
 import com.swekit.backend.Repository.SmsRepository;
 import org.springframework.stereotype.Service;
 import java.util.*;
@@ -16,24 +17,28 @@ public class SmsService {
     }
 
 
-    public List<Sms> allSms(){
-        return smsRepository.findAll();
+    public List<Sms> allSms(User user){
+        return smsRepository.findByUser(user);
     }
 
-    public void deleteSms(int id){
+    public void deleteSms(User user, int id){
         Optional<Sms> s = smsRepository.findById(id);
         Sms ss = s.orElse(null);
         if(ss == null){
             System.out.println("SMS does not exist");
             return;
         }
-        smsRepository.deleteById(id);
+
+        if(ss.getUser().getUserId() == user.getUserId()){
+            smsRepository.deleteById(id);
+        }
     }
 
 
 
-    public Sms addSms(SmsRequest smsRequest){
+    public Sms addSms(SmsRequest smsRequest, User user){
         Sms s = new Sms(smsRequest.getReceivedAt(), smsRequest.getSender(), smsRequest.getMessage());
+        s.setUser(user);
         return smsRepository.save(s);
     }
 }
